Fix a serious bug in virtual function call code generation
When outputs of a virtual function call were partially evaluated and
written to memory, and then that same virtual function call was compiled
a second time as part of another kernel that *also consumed the previously
stored call outputs*, Enoki-JIT generated nonsensical code (it tried to
copy the outputs once more from the virtual function call).

This triggered an LLVM compilation error (duplicate variable), and
undefined behavior in CUDA mode.