Seperate Optix pipeline and SBT configuration
It is important to decouple the OptiX SBT from the pipelines are the excepted lifetime of those two objects might be different. For instance in Mitsuba we reuse the OptiX pipelines across scenes to avoid recompiling those at every scene construction. However every scene has its own SBT.

With this patch Dr.Jit exposes jit_optix_configure_pipeline and jit_optix_configure_sbt, that will both return an index to a JIT variable responsible for the lifetime of those resources. Also an SBT is always dependent on a pipeline, and this is reflected in Dr.Jit as well.