Use consistent ordering of globals (callables & intrinsic declarations) in IR
When Dr.Jit is used in parallel programs, polymorphic classes may be
constructed in an arbitrary order. This is problematic since it causes
generated callables to shift around in the generated IR representation,
which breaks kernel caching. This commit refactors the assembly of
globals so that they are always uniquely ordered by their hash key.

This creates the following complication: when a virtual function call is
being compiled, we don't yet know the mapping "instance index ->
callable index". To determine the ordering of callables we must first
generate all of them (including those produced by other virtual function
calls) and then sort them.

This commit addresses the issue by introducing a new function
`jitc_vcall_upload()` that runs all the way at the end of kernel
compilation. At this point, the required information is readily
available, and the function can compute and upload the necessary data
onto the device.

The generation of globals was refactored to use a central buffer
('globals') and a sorted map ('globals_map') storing offsets into the
buffer. Once the kernel has been generated, the sorted entries are
copied from 'globals' to 'buffer'. The commit also includes some
superficial formatting changes to the generated IR.