Do a better job at detecting and adapting to the LLVM version
Dr.Jit must know the LLVM version to generate the right set of
intrinsics (which change between versions). To do this properly, it now
tries harder to detect the version version. When all fails, it uses a
fallback that resolves symbols known to only exist in specific LLVM
versions, which is good enough for inferring the major version.

The call to LLVMParseIRInContext() used to detect opaque pointer-capable
versions of LLVM printed an ugly error message onto the console in Linux
that could not be disabled. Furthermore, the LLVM docs state that opaque
pointers really shouldn't be turned on by default until LLVM 15. So this
commit changes the auto-detection into a version-based criterion.