Simplified memory allocator & LLVM thread scheduling
Dr.Jit previously used a intricate scheme for reusing asynchronous
memory allocations across threads and CUDA streams. Recent versions of
the system switched use a single stream per device, so this complexity
is no longer warranted.

The commit therefore removes most of the related logic from the
memory allocator.

Another part of this change adapts the LLVM backend to have a global
job queue (and not one per thread state) to be consistent with the CUDA
backend.