Unify `NodeType` and `VarKind` fields/enumerations
Following the previous sequence of commits, the Variable data structure
had a `kind` field that determined whether a variable represented one
of the following `VarKind` enumeration entries

- literal constant
- data (an evaluated variable)
- statement (legacy string-based IR)
- node (new abstract IR)

If it was of kind `VarKind::Node`, then there was another big `NodeType`
enumeration value stored in the `node` field accompanied by a
supplemental `uint32_t payload` field used to characterize the variable
in detail (for example, scatter nodes can have different
scatter-reduction types that are stored in the payload part).

This commit refactors the `Variable` data structure to use only a single
`VarKind` enumeration stored in the `kind` field as before. This
enumeration is the union of all previous possibilities. As a
consequence, the `payload` field was moved to the `literal` value and
grew from 32 to 64 bit. This will be useful for creating variables with
a pointer payload (e.g. a format string for `printf` nodes).