Limit the recursion depth when recording virtual function calls
Dr.Jit currently fails with an infinite recursion error when recording a
virtual function that is recursive. While this is simply problematic in
general with the recording approach, there is some use in supporting a
limited amount of recursion (an application in the Mitsuba context is,
e.g., a 2-level sensor or emitter).

This commit introduces a recursion depth limit to fix the problem.