Fix `gather` operations on results of other `gather` operations
With symbolic gather operations, invalid indices might have been pushed/propagated to previous gather nodes in the dependency tree which could produce out-of-bound memory reads if there was an other gather operation in the tree. This commit also propagates the mask throughout the tree to guarantee that no out-of-bound reads will be made.