Support new LLVM pass manager API
LLVM recently switched to a new pass manager API

- https://blog.llvm.org/posts/2021-03-26-the-new-pass-manager/
- https://llvm.org/docs/NewPassManager.html

Dr.Jit previously relied on the "legacy" pass manager API, and that
causes it to fail starting with LLVM 17 since the legacy API is no
longer exposed.

This commit resolves both APIs dynamically, and then uses the newer one
if it is available.