Fix race condition in ``jit_sync_thread()``
The LLVM thread synchronization releases the central Dr.Jit mutex and
calls ``task_wait_and_release()`` on the global variable ``jitc_task``.
This is a race condition because another thread can now also try to
synchronize and call ``task_wait_and_release()`` on the same task. This
will eventually lead to a double-release, which corrupts internal
nanothread data structures.

This is likely the root cause of various nanothread assertions that we
observed in the past.