Improve potentially buggy ``jitc_mkperm()`` task waiting logic
There have been some reports/suspicions by Miguel that a bug might be
lingering in ``jitc_mkperm()``, so I'm using this opportunity to
improve the task waiting/reference counting logic.

Instead of asking the ``jitc_submit_cpu()`` to not release a reference
to the preceding task, this commit uses ``task_retain`` to immediately
acquire a stable reference that it can wait on later.