Switch ``jitc_call_upload()`` and ``jitc_aggregate`` allocation method to avoid deadlocks
The LLVM versions of ``jitc_call_upload()`` and ``jit_aggregate()`` free
memory asynchronously by calling ``jit_free()`` from a nanothread task.
The trouble is that we sometimes wait for tasks to finish while holding
the Dr.Jit mutex, and this then leads to a deadlock when ``jitc_free()``
tries to acquire the lock.

This commit circumvents the issue by switching the allocations over to
``malloc()``. This is not expected to have any major performance
implications, since the allocations being done by these two code paths
are small.