Re-enable FP16 software conversion as fallback
It can be useful to compile Dr.Jit for really old processor
architectures (e.g. ``-march=sandybridge``) so that it works properly on
emulators and sanitization tools like ``valgrind`` that don't implement
all the fancy AVX2 and AVX512 instructions.

The latest half precision changes were problematic because they
triggered assertion failures on such architectures.

This commit adds a software fallback by Paul A. Tessier (@Phernost) that
is a highly crunched version of the code on the following StackOverflow
post: https://stackoverflow.com/a/3542975.

The fallback is *almost* equivalent to hardware FP16<->FP32 conversion.
The conversion perfectly matches for both normal and subnormal floating
point values including correct rounding. The only disagreement is
regarding how mantissa bits of NaN values should be treated. This is
arguably a very minor point, and hopefully nobody relies on them being
preserved or converted in any particular way.

A newly added tests checks the software fallback against the
hardware-accelerated conversion to ensure that this code path stays
functional in the future.