Better execution strategy for block-wise reduction
This commit significantly improves the implementation underlying
``dr.block_sum()``.  On CUDA, the reduction is now done using
cooperative warp operations and shared memory.

In some cases, it can also be useful to compile summation kernels
dynamically using global memory atomics. This is supported as well, by
specifying ``symbolic=1``.  If ``symbolic=-1``, the implementation
relies on a heuristic to choose a good strategy.