Adapted ``jit_var_inc_ref()`` signature
This commit changes the signature of ``jit_var_inc_ref(uint32_t index)``
to return the input ``index``. This is of course redundant, but it
shortens simultaneous assignments + refcounting operations:

```
variable = jit_var_inc_ref(index);
```

This is an improvement to the current pattern

```
variable = index;
jit_var_inc_ref(index);
```

which is found quite often. The change also makes AD and JIT-level
reference counting more similar to each other.

Finally, there is also a small potential code generation improvement
since the compiler does not have to stash the index on the stack if it
is needed after the call.