Workaround for OptiX linking issue in driver version R570+
When a two pipelines share the same set of direct callables, the driver
shares the compiled result. This caching contaminates the second
pipeline with globals from the first, which causes a linker issue when
the 'params' buffer has a different size. We work around the issue by
inserting a unique callable that prevents this optimization.