Added a new `VarType::BaseFloat` for Python type promotion (#132)
* Added a new VarType::BaseFloat for Python type promotion

This commit adds a dummy member VarType::BaseFloat to the VarType
enumeration. This type cannot be used for actual computation.

It is a placeholder to be used in situations where we haven't quite
decided what floating point precision to assign to a floating point
variable (E.g. a Python scalar float).

This allows us to compute things like Float16(1.0) + 2.0 without
the '2.0' being interpreted as Float32, causing an unwanted cast
to a higher precision.