JIT registry: avoid warnings due to ``jit_registry_clear()``
The function ``jit_registry_clear()`` clears the JIT registry, which is
useful for isolating tests from each other. However, if a cleared
pointer happens to be unregistered at some later point, this previously
generated a confusing warning message ("jit_registry_remove(ptr=...):
pointer is not registered!").

This commit changes the registry to mark cleared instances using a
special tag, in which case a later unregister call does not need to
issue a warning. Removing unknown pointer addresses is now treated as an
error.