Improved compilation of atomic packet scatter operations
This commit enables generating more efficient code for the ``PacketScatter``
IR node in CUDA.

- It uses the ``red.global.add.noftz.f16x2`` instruction when accumulating packets of F16 values.

- It potentially uses ``red.global.add.v2``, ``red.global.add.v4`` or
  even ``red.global.add.v8`` on sm_90+.

- Besides addition, minimum/maximum atomics are also supported depending
  on driver/GPU version.

This change also broadens the situations where packet operations can be
used on the CPU and GPU. For example, packet size 6 was not supported in
the past since it was not a power of two. Now, this is treated as 3
separate size-2 packets.